requirements.txt
astroid==2.3.3
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
colorama==0.4.1
cryptography==2.8
defusedxml==0.6.0
Django==2.2.3
django-cors-headers==3.0.2
django-rest-auth==0.9.5
django-rest-passwordreset==1.0.0
django-split-settings==1.0.0
django-webpack-loader==0.6.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangosaml2idp==0.5.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
nodeenv==1.3.3
Pillow==6.1.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
pyOpenSSL==19.0.0
pysaml2==4.9.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
